Move interpret() to separate file
