+++ smtp_handler/main.py
@@ -15,6 +15,7 @@ from datetime import datetime
 import pytz
 import django.db
 from browser.imap import *
+from browser.sandbox import interpret
 from imapclient import IMAPClient
 
 '''
