+++ smtp_handler/management/commands/mailbot.py
@@ -1,7 +1,8 @@
 from django.core.management.base import BaseCommand
 from schema.youps import ImapAccount
 from datetime import datetime
-from browser.imap import fetch_latest_email_id, authenticate, interpret
+from browser.imap import fetch_latest_email_id, authenticate
+from browser.sandbox import interpret
 from browser.models.mailbox import MailBox
 from imapclient import IMAPClient
 from engine.constants import msg_code
